Optimisations for post_processing and move over to OOP for codes in utils
