+++ environment.yaml
@@ -7,3 +7,4 @@ dependencies:
   - pandas
   - matplotlib
   - geographiclib
+  - bs4
